modularize get playlist
